# Improve filter performance :