--- a/common/datavalues/src/columns/common.rs
+++ b/common/datavalues/src/columns/common.rs
@@ -1,6 +1,3 @@
-// Copyright 2020-2021 The Datafuse Authors.
-//
-// SPDX-License-Identifier: Apache-2.0.
 
 use common_arrow::arrow::array::Array;
 use common_arrow::arrow::compute::concat;