--- a/fusequery/query/src/pipelines/transforms/transform_filter.rs
+++ b/fusequery/query/src/pipelines/transforms/transform_filter.rs
@@ -21,6 +21,7 @@ use crate::pipelines::processors::EmptyProcessor;
 use crate::pipelines::processors::Processor;
 use crate::pipelines::transforms::ExpressionExecutor;
 
+
 pub struct FilterTransform {
     schema: DataSchemaRef,
     input: Arc<dyn Processor>,
@@ -93,6 +94,14 @@ impl Processor for FilterTransform {
             // Downcast to boolean array
             let filter_array = filter_array.bool()?.downcast_ref();
             // Convert to arrow record_batch
+
+            let mut  filter_exit_true =  filter_array.values().chunks::<u64>();
+
+            if  filter_exit_true.find( |p| p.count_ones() as usize > 0_usize ).is_none() &&
+                filter_exit_true.remainder_iter().find(  |p| *p ).is_none(){
+                return  Ok(DataBlock::empty())
+            }
+
             let batch = block.try_into()?;
             let batch = arrow::compute::filter::filter_record_batch(&batch, filter_array)?;